starter-kit/requirements.txt
numpy # Don't remove
Pillow # Don't remove